Remove and register changed to only hash-meta and hash

+++ AuthService/src/main/kotlin/dev/easypass/auth/datstore/CouchDBConnectionProvider.kt
@@ -52,6 +52,7 @@ class CouchDBConnectionProvider(private val properties: Properties) {
      * @param dbname: the name of the database
      */
     fun deleteCouchDbDatabase(dbname: String) {
-        createCouchDbInstance().deleteDatabase(dbname)
+        if(createCouchDbInstance().checkIfDbExists(dbname))
+            createCouchDbInstance().deleteDatabase(dbname)
     }
 }
